Moves the logic that generates our User Agent header into its own class.
